iSNAP-110: Added the capability to POST a form and accept the changes.
Re-submitting because it did not work the first time.
snapmanager.cgi displays a large WARNING in case you use it on a non-secure connection (i.e. no SSL).
Restored the messenger of the snapmanager.cgi process. Now we can send a new variable value (MODIFYSETTING).
Changed the MANAGEREPLY into MANAGERACKNOWLEDGE.
Show the state in the host display screen of snapmanager.cgi.
Fixed the generation of the form so it includes the query string.
Added attributes to the form so it comes out as UTF-8 and a POST (on top of keeping the query string.)
Prevent the display of a form if that variable is marked as "modified".
Fixed the ownership and permissions of the database files because we now want the snapmanager.cgi to be able to write to it (i.e. so that way we can mark a field as "modified",)
Fixed the form hidden fields, they had no name. Added the hostname too.
Added a test to snapinit that was using CONTENT_TYPE without verifying that the variable existed (i.e. it could be a nullptr.)
Fixed the load_plugins() so it can get called more than once.
Added an auto_close RAII class in the server_status::write() function.
Moved the -DDEBUG and -D_DEBUG and -DNDEBUG and -DCONTROLLED_VARS_DEBUG definitions to the cmake/Modules where they should have been.
As a side effect, updated the controlled_vars tests so if CONTROLLED_VARS_DEBUG is already set, it does not try to set it again.
Deleted the snapserver.logrotate because all of these entries are actually in their respective service or under snapbase. Having two copies can create problems.
SNAP-361: Fixed the DISCONNECTED overload and used HANGUP instead. snaplock understands both messages now.
Various cleanups.