SNAP-347: Work in progress. Refactoring to use events.
I added a new queue of function objects, and the idea is to push your
function onto the queue and let service::process_timeout() handle it.
However, the current design is fraught with lots of status variables,
and functional side effects. This is going to take some doing to
refactor, so I've created a branch to do it in.