SNAP-110: Added support to gather the bundle files. (i.e. files describing what can be installed.)
Fixed bug calling replace_configuration_value(), it now returns a bool, I was still expecting an int.
Fixed the POST value as it may need decoding (for :, /, %, etc.)
Changed the data_path variable to point one directory up. We now append cluster-data and bundles to that path.
Started support to use STL instead of Qt strings (i.e. added split(), join(), and cstring_reverse to iterate char * in reverse).
Added the missing button implementation: it reads the string, make sure it is correct (not tainted) and then it passes down to the "apply settings" function.
Also I set the server and server names in the MODIFYSETTINGS message according to the button name.
Added a function to extract the basename in a path from an std::string object.
Added a description widget to the form (in part because otherwise we get nothing in there...) It's just outputing basic HTML though (no form elements).
Added the Install, Uninstall, and Reboot buttons.
Finished implementation of the installer: it can install, uninstall (purge), and upgrade (update, upgrade, dist-upgrade, autoremove).
Implemented a function to check whether a package is installed.
The self plugin also adds a field to let the administrator know whether a Reboot of a machine is required.
Various clean ups.