SNAP-347: Final features put in refactor.
Now the service using a queuing mechanism to handle service status
changes. This is a superior design to just relying on status variables.
It allows is to queue events based on the current status much easier,
and react to changes without having raw loops, huge switch statements
and massive if/else blocks.

As far as features go, the ability to stop prereq services is now
implemented, as well as requesting their restart when the service is
running again.

I keep the timer running all of the time. It just silently returns if
the queue is empty.