SNAP-168: Fixed a bug in the processing of a message, a remote connection may be open on a service_connection (i.e. connection we received, opposed to connections we make...) and that was not taken in account when specifically sending a message to that server.
SNAP-110: Fixed the destination host, I used 'f_server_name' in snapmanager.cgi, when the correct parameter was 'host'.

Added the code to actually initiate a reboot on a computer.
Fixed the name of the service sending REGISTERing snapmanager.cgi. It needs to be a unique name so we add PID to it.
snapbase creates the all.log file so it can give it the correct owner and group (otherwise it can end up being root:root because it could get created by snapinit.)
Fixed the properties, by default they are meant to have DEBUG and not TRACE in the "other sample with console".
Various clean ups.