SNAP-110: Added function count_packages_that_can_be_updated() to know whether we have to upgrade that computer.
Added 3 new variables to support checking for upgrades.
Added the reboot_required variable so if it moves we can easilly fix the path (especially on different OSes).
Added an UPGRADE button and its implementation.
Added a function to reset the apt-check results after an upgrade or an install call.
Fixed the shutdown call to a COMMAND since we do not care about the output.