SNAP-347: Finally tuned the algorithm properly to stop and start prereqs.
When a service dies, you have to make sure the services which depend on
them (what I call prereqs) be shutdown, but only after the service
cannot be restarted.

If the service suddenly works again, and snapinit is able to start it
back up, then you have to also bring the prereqs back up. The prereqs
will wait endlessly until what they depend on start in that case. Then
they will be restarted.