SNAP-110: Add support for a set of fields to install a bundle (that way we can immediately install it properly--otherwise there are situations where installation would be cumbersome.)
Added a preliminary bundle to install OpenVPN as a server.
Added support for preinst and postint scripts.
Fixed the gathering of the name and description of a bundle with firstChildElement() instead of elementsByTagName().
Added an Apache2 configuration file to prevent httpoxy.
Added code in snap.cpp to avoid ever transferring the HTTP_PROXY variable to snapserver/snap_child.