SNAP-110: The installer "install" command does not do an upgrade anymore. We force the user to first run an upgrade on its own, then we allow installations.
By separating the "install" from the "upgrade", we could keep the "install" blocking whereas the new version of the upgrade has to fork() now.
Added a lock file that locks the install and/or upgrade while it is happening. Further install and upgrade calls are totally ignored (i.e. we should probably queue them instead?)
Added a path to the lock directory to use to make sure we only upgrade once at a time.
Added an STL based lockfile object (in parallel to our existing qlockfile which works with QFile.)