SNAP-110: For the snapupgrader to not get killed we MUST change the snapinit.service to use KillMethod=process instead of the default (control-group).
Added full error handling of the return value of starting the snapupgrader command. This gives us a better view in case something happens.
Added a call to setsid() to put snapupgrader in its own session, although that makes no difference to systemd, it makes it clear that it is not part of the other running daemons (if we were to search for all fork()'ed children with the same SID...)