SNAP-110: Added all the bundle files that I can think of at this point. These have no preinst/postinst yet, but they should all be there. Also their list of packages may be erroneous.
Added the corresponding install() and snap_validate_xml() rules to cmake.
Moved the .xsd files under their own directory (we now have 2 of those).
Added support to the snap_validate_xml() to search for .xsd file under the xsd directory instead of the dtd directory.
Added the snapserver-core-plugins to snapbackup since snapserver may not be installed on that computer and thus the plugins would be missing.
Fixed the clear-cache script in snapdb so it actually does something useful for users.
Various cleanups.