SNAP-110: added the libxml2-utils to snapbackup now that it needs to run xmllint.
Added my showfw script to snapfirewall so that way we can easily check the firewall status without having to remember the exact iptables syntax.
Added the antivirus bundle to the core plugins. That creates a problem in the implementation of the version signal. i.e. the computer that executes that callback will not be the one that has clamav installed.
As a result, I moved the clamav from Depends: to Suggests:.