SNAP-110: Added the firewall plugin to allow for "enable/disable" and "recovery" amount editing from snapmanager.cgi.
Added the snapmanagercgi project as an include path to the snapfirewall because we now have a plugin there that depends on it.
Stengthen the XSD source path by adding it to the list of searched paths.
Cleaned up various Suggests and Recommends. Especially, we do not want to auto-install anything without using the corresponding bundle now.
Removed various flag definitions from snapfirewall main CMakeLists.txt. They did not do anything wrong but we should have all of those in the snap main place defining such.
Added some new dependencies to the snapfirewall. Some of which became required because of the plugin.
Added the RELOADCONFIG message in snapinit. It does nothing unless snapinit runs as root (so programmers do not get a surprise).
Added a cmake file for snapmanagercgi so one can develop plugins in their own environment instead of all under snapmanagercgi.
Just in case, added the --no-install-recommends to the apt-get install command.
Made sure that the snapmanagercgi header files from the library were compatible with snapfirewall's plugin.
Various cleanups.