SNAP-347: Refactored the states so that we don't rely on statuses.
The states should be far more flow-through, and less likely to stall.
I've improved them significantly, and also removed the old statuses. Now
when you called failed(), is_stopping(), has_stopped(), etc, we now look
at the f_current_state variable, which is now set after the state is
puleld off of the queue, and before it is run.