SNAP-347: Added paused/pausing states and took snapcommunicator out of general services list.
Now when a process dies, it pauses the prereqs, instead of failing
them. The "pause" is a special state where the process is killed, but
kept in the communicator list, because we know at some point, the
process might come back up. In which case, the services are "unpaused"
through the starting/started states.

The instance for snapcommunicator is no longer stored in the snap_init::f_services
list. This will prevent services from "pausing" it when a service is
marked "failed." In fact, it is no longer needed as a dependency, since
snap_init makes sure it is started first before all other services.