SNAP-109: Added code to send an RUSAGE message whenever a snap_child object exits.
Updated the snapwatchdog to accept the RUSAGE message.
Changed the f_processes variable to actually be a vector. We may have many RUSAGE coming in very quickly and one tick also.
Fixed the watchdog_server::process_sigchld() function: we have to check all children that have exited before we return.
Changed the algorithm slightly to take in account the fact that f_processes is now a vector.
Allow for the data_path parameter to be empty, in which case no external .xml file get created.
Fixed the saving of the resulting statistics in the database. I would not use the 'value' variable with its TTL!!! That explains why that table would grow forever!!!
Removed the controlled_vars<>() from the snapwatchdog header as I was at it. Using C++11 initialization instead.
Added a pointer_t to the snap_cassandra class.