SNAP-347: Remodeled the state machine to actually properly handle the quitting part in all cases. Especially I created 4 state machines instead of just one big monster which was very difficult to deal with.
Added a process class that handles the process status (stopped, unregistered, registered, error). Moved all the service variables that are specific to the process inside the process instead (i.e. coredump size, nice value, etc.)
Added a clear state machine declaration in a comment at the top: disabled, ready, paused, go down, stopping.
Also added a sub-state machine for the stop process which involves multiple states: idle, stop, terminate, kill, abort process.
Re-added dependencies and actually we now have a snapinit service (an actual object) and we can that way create a dependency from snapinit on to snapcommunicator (i.e. we first need to get the connection and then we can move on.)
Added the --tree command line option to be able to generate a graph of the service tree.
Fixed a few things in regard to --list which was not working 100% as expected.
Removed the second exists() test, it is done in the process::set_commands() now and that's more than enough.
Added the sender's PID to the SAFE message. That way we can easily find that service in snapinit.
Added support for weak dependencies. If that service is not installed on that machine, that dependency is ignored.
Put snapinit definitions within its own namespace.
Changed f_wait_interval to clamp it on read so we do not have to do it each time we use the value.
Renamed the snapinit.xsd file as service.xsd which makes more sense.
Updated the service-README.txt file with the newest capabilities and fixed up some existing entries to better match the corresponding implementation.
Extracted the common namespace and put it in its own .cpp/.h files.
SNAP-351: Moved the nice value support from the snapbackend to snapinit, that way any daemon can be given a low priority, not just the backends.
Also remove the nice=... value from snapcommunicator and use a <nice> tag instead.
SNAP-340: The newer version of snapinit knows whether a process was registered (i.e. STATUS up was received) or not. So it immediately sends a SIGTERM "as expected".
Commented out the use of the NEWSERVICE message since no other process uses it and it is not as good as the STATUS message.
Various clean ups.