SNAP-347: Actually thought that snapinit should be started in /var/lib/snapwebsites ("home" for snapwebsites) so I added a chdir().
Updated the start and stop scripts accordingly since the chdir() forces me to have full path for the --binary-path option.
Stop erasing services from f_service_list. This is a problem when exiting because we may be in another std::for_each() or similar loop which then crashes because the list was modified under our feet.
Changed the "list is empty" test to check that all the pointers in the list are nullptr.
Changed the weak dependency definition using a type instead of weak="weak". The type can be type="weak" or type="strong".
Save the disabled status so the --list and --tree can give us the information.
The --tree also could display the current tree and show paused elements in red.
Added error strings to the snap_thread implementation.
Various clean ups.