SNAP-69: Removed the install() target, the AddDoxygenTarget() macro already takes care of that!
Doxygen was most marked as REQUIRED in many CMakeLists.txt files.
Removed several install() target in advgetopt.
Removed the find_package() of Doxygen in a couple of projects where it is not used.
Protected the add_dependecies(), and a few other functions in csspp in case Doxygen is not installed.
Added license to libQtSerialization doc/CMakeLists.txt (it was missing.)
Removed the ip4 example in firewall since we install our complete firewall rules anyway.