SNAP-347: Fixed a bug in the common.cpp--it would call the snap_init->exit() function and delete the PID file if a process child had a problem.
Added a disconnect function that includes sending an UNREGISTER message to the snapcommunicator from snapbackend.
Noticed that the CRON snapbackend would be killed after a while (paused, really.) I changed the process state machine, we now receive the exit code which allows us to either have an error or just "the process died normally". This way the error counter does not get incremented when snapbackend exits normally.
A paused process does not anymore send a GOINGDOWN to a dependency which is a weak link.
Added a sanity check to make sure "snapcommunicator" appears first and then "snapinit". It is important for these two to have very low priority in that order.