SNAP-390: New version of the script which does everything all in one.
This new version goes through all of the projects, finds the debian
folders, bumps the changelog, then updates all of the projects in the
tree based on which depends on what. The control files will reflect all
of the new versions they depend on.

This is a work in progress--this work is untested.