SNAP-390: One must specify the search cache.
Now for the perl scripts which search for dependencies, the command line
arguments have changed. Now instead of silently creating a cache file in
the /tmp area, you must explicitly specify it on the command line. I've
added a new script, called SnapMakeDepsCache.pl which you use to build
the cache the first time. Then both SnapBuildIncDeps.pl and
SnapFindDeps.pl both use this cache. They no longer accept the root
folder as an option on the command line.