SNAP-110: Added support to auto-setup apache2 when installing snapmanager.cgi
Added the snapmanager-apache2.conf, enable it, and make sure that the "cgi" module is also enabled.
Added an index.html so people get some documentation if they just try to access their server directly.
Especially, the index tells people to edit their snapmanager.conf file to add their IP address in there.
Changed the dependendy list of the snapmanager.cgi package, it actually requires the snapmanagerdaemon package.
Moved the bundles footer to the snapmanagerbuild package.
Put the index.html under the share folder as well so administrators can have a copy of the original handy.
Updated the documentation in the snap-httpoxy.conf since it applies to snapmanager.cgi too.
Made sure to restart apache2 when enabling the snap-httpoxy configuration file.
Made snap.cgi depend on apache2. There is no reason not to since without Apache snap.cgi is totally useless.
Extended the initialization of the snapcgi.log to the secure directory and removed the write to the group permission. No need.