SNAP-110: Made a change to snapinit so on loading we verify the server name validity. '-' become '_', '.' mark the end of the FQDN, and uppercase is forced to lowercase (FQDN are case insensitive).
Applied a fix so if no hostname can be found, snapinit exit with 0 instead of 1 to avoid having systemctl restart the process over and over again.
Made snap_communicator::verify_name() a static and public functions so we can verify a name from anywhere outside of the snapcommunicator library.
Changed the RestartSec parameter of snapinit.service to 10 seconds. That way if it crashes constantly we at least don't burn the processors too much.
Commented out the clients=... variable in the "snapmanagercgi.postinst" because an admin who edits the main snapmanagercgi.const could otherwise get a surprise.