SNAP-110: Added a Refresh button, apache2 plugin to edit the ServerName, and fixed the ssh filename.
Added "Refresh" button which asks to all snapmanagerdaemons to re-check their computer status immediately.
The Refresh will always appear first. It is not a field in the list of fields from the status file, so it is handled specially.
Added a WAKEUP message so we can ask the snapmanagerdaemon status thread to immediately re-check the status.
Along the refresh addition, I added the necessary code to be able to use UDP signals from snapmanager.cgi (namely, getting the snapcommunicator_signal variable info.)
Added a file_content.cpp/h to load/write entire file at once.
Changed the behavior of the affected_services set of strings: a plugin can now add a name which is not a snap server, but they have to the handle that name in their on_handle_affected_services() function.
Added the definition of the "snapcommunicator_signal" variable to the snapserver.conf which makes use of it in case it wants to send a PING message.
The snapmanager.cgi now includes a "text" widget which uses a textarea object without wrapping text (since 99% of the time this will be for configuration files.)
Added a search function in the installer so we can search parameters in various types of configuration files. Used for the Apache2 .conf files for now.
Also included a "snapcommunicator_signal" variable in the snapmanager.conf file, that means we now have THREE copies.
Added the apache2 plugin to edit the ServerName. You can also reset it back to the default (i.e. commented out). Tested on my system against one of the files and it works very nicely.
Changed the affected_services vector in a set since we do not want duplicates (and a map is not as good in this situation).
Fixed the SSH plugin so it makes use of the "authorized_keys" file and not "id_rsa.pub". It will show all the keys and you can add a new one and remove existing keys.
Various clean ups.