SNAP-110: Removed the `rm -f .../service-*.xml` commands from the postrm. Instead we updated snapinit to be tolerant in this case.
Made postfix a dependency of snapbounce because really snapbounce is useless without postfix.
Strengthen the script for postfix initialization, although we need to run that script to make sure it works.
Fixed up the frontend bundle. Since snapcgi depends on apache2, we do not need it in the bundle and any package initialization is now done there instead. We still initialize the domain name and webmaster email address from the bundle script.