SNAP-110: Updated the backup bundle to include a warning about the IP address that needs change once installed or it won't backup anything.
Added a HOST variable and the --host option on the snapbackup command line in the cron.daily file.
Added a chmod 600 on each backup so that way they get protected from preying eyes.
Added a bungle module for snapbackup so one can change the IP address of the host (i.e. Cassandra's).
Various cleanups.