SNAP-110: Fixed the bundle tools (name was still "ssh")
Removed "screen" from the bundle tools, because it is actually a default tool installed by the OS.
Added code to the ssh bundle to make sure password login and root login are both turned off.
Added a warning in the index.html about the potential conflict between the snap.cgi and snapmanager.cgi packages.