SNAP-110: Moved the creation of the snapserver.log file to snapbase because it may be used by snapserver or snapbackend.
Fixed the sending of wrong messages to services that do not otherwise expect them. I was keeping a list of connections interested by the message, but that included local services, which was wrong.
Added a lock file to the snapdbproxy since it uses a thread and I've seen problems in that file before. Somehow, only processes with multiple threads have shown some problems like this.