SNAP-403: Now starts openvpn service correctly.
We needed to be able to create the server.conf file, in addition to the
server certificate. This is now done. Additionally, if the dh2048.pem
file already exists, it won't be recreated (that is because it takes a
long time).

Side note: there actually is no need to specify the "Common Name", so I
took it back out.