SNAP-403: New vpn plugin handles client cert generation.
The idea is that when you configure a server for the first time
(postinst), the server keys will be created, but so will each client.
Right now, this only works one time through. At some point, I need to
add support to handle future client-adding.

Also this code is completely untested and uncompiled.