SNAP-410: Removing snapinit and replacing with systemd .service files. (snapinit does not build anymore, however, we still have the folder for now until we get things to work as expected.)
Also added a .timer file for the CRON snapbackend tool. We may still want to consider transforming that tool to make use of fork() and children on its own (like the other snapbackend actions).
Reimplemented the snap_config.cpp/h to better allow us to load configuration files from any .conf.'
Made a master snap_configurations class to cache the configuration files that ever get loaded.
The snap_config class is now a wrapper of the new snap_configurations class.
As a result of that change, removed the --connext and --snapdbproxy references (there may still be one or two, but most are already gone).
Changed the --config command line option to accept a path to the configuration files, but not the filename. So you cannot rename the files anymore.
The snap_config_file class now forces the configuration file name extension to be .conf.
Moved the --debug command line option to the corresponding .conf file (i.e. this daemon .conf, that variable is not shared.) Not yet reimplemented in each daemon since it is not used much anyway.
Replaced the get_server_name() of the server class in the library, with the function we had in the snapinit daemon.
Removed the "snapcommunicator_local" variable from the snapserver.conf since we now read that parameter directly from the snapcommunicator configuration file.
Noticed that the -Wint-to-pointer-cast was not required anymore with BIO_set_conn_int_port().
Removed all the service files from the packages (they won't get installed.) The files are still here for reference. We will remove them once the new scheme is at least mostly working.
Also removed the installation of the service.xsd file.
Moved my start and stop scripts from snapinit/scripts to snapbase/scripts. This is only for developers.
Fixed the snapbounce .conf file -- we are not using it, but it now references the snapserver.conf which exists, at least...
Updated the snapmanagercgi plugins to make use of the new snap_config interface.
Moved the table_schema_path to the snapdbproxy.conf file. Tools in that service will be using it.
Added the snapfirewall to the new scheme: it now reacts to the CASSANDRAREADY and NOCASSANDRA messages. It sends the CASSANDRASTATUS message once to make sure it gets notified as expected.
Now snapfirewall sends the FIREWALLUP message once it was able to setup the firewall entirely. We'd need to keep that information and answer a FIREWALLSTATUS message that way too.
Moved the server_name variable from snapinit (deprecated) to snapcommunicator.
Updated various control files so they do not reference snapinit.
Updated the server class implementation as the default f_parameters cannot be defined early anymore. They have to use the set_parameter_default() after the file got loaded. (If not loaded yet, it will load it on that call.)
Added a disconnect() function, when a daemon receives the NOCASSANDRA message from snapdbproxy, this function is expected to be called.
Finally removed the init_context() function. Instead we have to make sure to run snapmanager once (to be moved to snapmanager.cgi / snapmanagerdaemon.)
Added a snap_cassandra::get_table() instead of the create_table(). The creation of the tables is to be done by the snapcreatetables tool now.
Changed the snapwebsites.h.in for a version.h.in in the snapwebsites library.
Updated the log.h(.cpp) to support the snap_config_parameter_ref so that way we can still support QString. Once QString is out of snap, we won't need that extra support.
Moved the create_tables() function from the server class to the snap_cassandra class. More sensible.
Prevent snaplayout from creating the layout table. Instead you have to run snapcreatetables as expected.
Various clean ups.