SNAP-410: Fixed the version changes to the snapwatchdog daemon.
Fixed the connection to snapcommunicator, it was using the wrong IP address making the snapwatchdog look like a remote snapcommunicator.
Fixed the NOCASSANDRA implementation, we needed to clear the snapdbproxy address and port.
Prevent running the children if we cannot connect to the database (this is not correct, but the current implementation forces the connection to the database...)
Added a message to see the current list of connections available in snapcommunicator (for debug purposes).
Removed another controlled_vars reference.