SNAP-410: Fixed problem with finding snapwatchdog.h
In the src folder, snapwatchdogserver was not looking for the header file
"snapwatchdog.h" in the right place (under ${CMAKE_SOURCE_DIR}/lib). So
when we build a package in a pbuilder chroot, it would fail to find the
header. It worked just fine on the developer systems, because likely, an
older version of snapwatchdog.h was sitting in the
dist/include/snapwatchdog folder. This would give a "false positive,"
and make it look like the project was compiling and picking up the
correct version of snapwatchdog.h.