SNAP-409: Changing the prepare_cassandra() in a check_cassandra() in the server class implementation (a la snapwatchdog).
Removed the --add-host since we do not need such for the snaplock tool to work (it gets our server name as expected on its own.)
Removed the --prepare-cassandra option which was just replaced by the snapcreatetables tool until it gets in snapmanager.cgi.
snapwatchdog now makes use of the new check_cassandra() function from the server instead of using its own.