SNAP-406: Added support for the NOCASSANDRA and CASSANDRAREADY messages in the snapserver.
If snapserver cannot connect to the Cassandra database, it won't create children to handle connections. Instead it immediately fails with a 503 error.
Applied a fix to the snap.cgi that would ignore premature errors from the snapserver (such as that new 503 error).
The snap.cgi may reply with an early server error or its internal 504 error as appropriate.
Fixed the snapwatchdog as it was never sending the CASSANDRASTATUS signal, which means it could miss the first NOCASSANDRA and CASSANDRAREADY.
Various clean ups.