Fixed problem with changelog version parser.
When reading a version which does not have four parts
(major.minor.patch.build), the original code would silently break, and
return wrong version values. This caused the Doxygen make code to fail
in a strange way.

I've built in a safeguard: if the version is not the right format, a
FATAL_ERROR occurs and CMake processing ends.