SNAP-406: Added a `close()` of the socket on any error in the snapdbproxy connection thread. Otherwise the client sits there until another connection arrives, which could take forever.
added a few TRACE in the snapdbproxy connection thread so we see when a thread enters and exits (we'd need to show the thread ID though...)
Added a version.h.in to snapdbproxy.
Display the version of snapdbproxy in the log on startup.
Display the correct version when using the --version command line option.