SNAP-417: Added a mutex to the snap_config implementation so we can safely use it in a multi-thread environment (like snapmanagerdaemon).
Removed the need for controlled_vars in snap_thread, however, I did not yet remove the #include as they are probably used by other headers.
Fixed the snapcommunicator so it will ask itself to restart on changes, instead of snapinit.
Implemented the RELOADCONFIG in snapcommunicator.
Commented out code in snapbackend and snapdbproxy plugins because that code is not currently supported.