SNAP-110: Fixed the file that the awk script of the snapfirewwall bundle changes to /etc/network/interfaces instead of ~/tmp/d.txt (i.e. my test file!)
Fixed the firewall rules: we want to allow public and private connections in bad_tcp_packets.
Fixed the firewall rules: if the public and private interfaces both use the same IP address, it requires that the REJECT of public IPs appear after we allow private interface IPs.
Fixed the test to know whether all the variables were properly changed from BUNDLE_... to an actual IP or interface name.
Fixed the firewall bundle so it says "space separated" instead of "space separator".