SNAP-110: Added code to send the load average only if it changed since last time (TBD: we may also want to send it at least once a minute as a watchdog?).
Fixed the help in the user snapmanager.conf which we create in postinst.
Make use of a variabel so if we want to change the filename, it's done in one place.
Added such a sample user snapcommunicator.conf since, except for the first installation, the user is likely to have to edit that file.