SNAP-110: Fixed the snapfirewall, it would block all IPv6 traffic by default, even though our computers may use IPv6 on lo which should never be blocked.
Added the NEWREMOTECONNECTION event on an ACCEPT too. After all, it is new whether we receive a CONNECT or an ACCEPT.
Removed the "internal" return from the dbproxy.cpp plugin. That would prevent adding further fields after that try/catch block.
Also changed the ERROR in a WARNING because people installing anything that requires snapdbproxy will end up with that "error" otherwise...