SNAP-110: Removed spurious warning from snap_cassandra.cpp get_table(). That warning was for the create_table() function only.
Fixed the name of the data center when creating the context. It has to use the same case as the one defined in the file, which is "dc1" and not "DC1".
Found the error about the QString missing arg(). I'm not too sure how I searched, but somehow I found the wrong message yesterday. Today I got to the right message and fixed it.