SNAP-110: Fixed the snapdbproxy implementation so it does not close the same socket twice (because the second time it could be a new socket!!!)
Commented out log statement in snap.cgi at a location where the logger is likely not initialized yet.
Added comments in the snapcgi/tests/request.sh about how to run the script.
Applied a fix to the ip4tables.rules as the source of 4004 and 9042 can be set to @ADMIN_IPS@ by default.
Various clean ups.