SNAP-110: At this time, the time tick may occur when CASSANDRAREADY was not yet received. When that happens, it looks like the database is not ready and thus the tick for the CRON task would fail.
Fixed the service files as we do not want the STOP message to ever get cached.
Added a TTL to the messages added to the snapcommunicator cache. By default messages are kept for 1 minute. One can change the TTL to any number of seconds (add_parameter("cache", "ttl=20");).
Added a way to mark a message non-cacheable with add_parameter("cache", "no").
Found a way to avoid one copy of the cached messages.
Removed snapinit specific code from snapcommunicator.
Added the log_all entry to the snapwatchdog.properties.