SNAP-110: Removed the "snapmanager_frontend" field management from snapmanagercgi plugin. It's not really working as expected at this point.
Moved the dependency on snapdbproxy from the libsnapwatchdog-dev to the libsnapwatchdog package where it belongs.
Removed the "VPN Server IP Address" from the VPN bundle. Not used/needed.
Added important note to bundle about which IPs are to be entered here.
Fixed the .install of the snapwatchdog as it places 3 sub-directory under usr/shared/snapwebsites and we place each one in a different package.
Various clean ups.