SNAP-110: Allow the server name to be set to "*" so we can send to a specific service to all servers (needed for snapcommunicator itself).
Applied some fixes to the FORGET implementation, although I could not prove that it was fully working yet.
Broadcasting works by testing whether the service is set to ".", "?", or "*" and if not one of these, then the server name is used.
Added a snapserver manager plugin so we can activate that server from snapmanager.cgi.