SNAP-110: Implemented the RELOADCONFIG in the snapserver.
Fixed the path to the system files (it was /lib/service instead of /lib/systemd).
Fixed the variable name to the timer OnUnitActiveSec, it's part of the Timer section, not Service.
Fixed the affected services on installation, use the actual content of the tag instead of "install_values".
Removed the controlled_vars from the snapserver header.