SNAP-110: Completely removed the f_started variable and start() function. Adjusted code accordingly. Readded the 1 sec. delay between connections.
Changed the remote communicator connections to be enabled by default. It now starts right away, no matter what.
Remove failed connection from the gossip and smaller ips lists.
Properly reset the timeout delay on a successful connection to 5 min. It could be changed to 15 min. because of a shutdown.
Various clean ups.