SNAP-110: snapwatchdog CPU plugin uses the procps, so I re-added the file in the control file.
Moved a couple of tests from snapserver to libsnapwebsites. These are not really specific to the library, but they were used to determine things that eventually got developed in the library. The pmon.c is also in this directory, although I have not added a project in CMakeLists.txt yet.
Removed another controlled_vars dependency.