SNAP-110: Added two fields to edit the After= and Require= parameters of the snapcommunicator.service file because without the tun0 a reboot may attempt to start those services too soon and they will fail.
Added some of the hard coded strings to the get_name() function.
Fixed the backend plugin so it uses snap::process instead of system() so we get info about what's going on.
Moved the header files of the snapmanagercgi/lib under a sub-directory named snapmanager (As it gets installed in such) so that way we can have #include "snapmanager/..." everywhere as expected.
Added support for sections in the replace_configuration_value() so we can replace and add values within a section. The "only" odd thing is it will add "in the wrong place" (i.e. after the empty line.)