SNAP-110: Added a command line tool so we can install a website without the need to have snapmanager (since most of the code is in the library anyway.)
Moved all the libsnapwebsites headers in a sub-directory named snapwebsites. Fixed all the corresponding #include so they start with "snapwebsites/...".
Fixed the SnapWebsitesConfig.cmake to search for snapwebsites/snapwebsites.h (no prefix since we force include that prefix in the header we are searching for.
Removed the -I option that included the prefix that way we always have to use it in the #include command.
Fixed the configuration of the version.h.in since now it is in the sub-directory. All those changes makes sure we are including the snapwebsites version.h file and not another.
Fixed up some #include lists as a result, removed many commented out entries too.
Fixed the initialization of the "listen" parameter in the server. The default must be 127.0.0.1 and not 0.0.0.0.
Removed the controlled_vars from http_strings.h, plugins.h, snap_initialize_website.h, snapdb.h.
Fixed the #include for inet.h in inet_helpers. It needed to use < and > and not ".
Added some #include <snapwebsites/poison.h> to make sure we do not use dangerous C functions.
Removed inclusion of snap_tests.cpp from test_snap_cgi since it is not used there yet.
Removed the snap_tests.h and snap_tests.cpp duplicates found in the snapcommunicator. These are now only in snapbase.
Moved our #include from the list of system includes to standard includes.
Moved the version.h.in from snapmanagercgi/lib to the snapmanagercgi/lib/snapmanager sub-directory.