SNAP-110: Fixed snapcommunicator, the IP addresses read from the cache need to be added to the list of remote communicators to listen to at the time they are read.
Reduced the amount of time we wait before trying to reconnect to Cassandra to just 1 minute max.
Bumped version of snapcommunicator so I can test again now.
Added a warning in the `droptables.sh` output about the fact that the drop is not going to be known by `snapdbproxy` unless restarted. The create tables are known, but the drop bypasses the `snapdbproxy` at this point.