SNAP-440: Added the start to the messaging for SSL keys.
Two new messages have been added: CASSANDRAKEYS and CASSANDRAKEY. The
former is a request made from snapdbproxy which asks the "cassandra"
snapmanagerd plugin to send the available SSL public keys from the
Cassandra nodes on the network that snapcommunicator knows about.

The latter message is what the plugins send, and snapdbproxy stores the
keys locally. If connected, the session is disconnected, and the local
SSL keys are added to the session, and then we reconnect.